+++ Craftsman/Builders/Projects/WebApiCsProjBuilder.cs
@@ -13,14 +13,18 @@ public class WebApiCsProjBuilder
         _utilities = utilities;
     }
 
-    public void CreateWebApiCsProj(string solutionDirectory, string projectBaseName, DbProvider dbProvider)
+    public void CreateWebApiCsProj(string solutionDirectory, string projectBaseName, DbProvider dbProvider, bool useCustomErrorHandler)
     {
         var classPath = ClassPathHelper.WebApiProjectClassPath(solutionDirectory, projectBaseName);
-        _utilities.CreateFile(classPath, GetWebApiCsProjFileText(dbProvider));
+        _utilities.CreateFile(classPath, GetWebApiCsProjFileText(dbProvider, useCustomErrorHandler));
     }
 
-    public static string GetWebApiCsProjFileText(DbProvider dbProvider)
+    public static string GetWebApiCsProjFileText(DbProvider dbProvider, bool useCustomErrorHandler)
     {
+        var errorPackages = "";
+        if (!useCustomErrorHandler)
+            errorPackages = $@"{Environment.NewLine}    <PackageReference Include=""Hellang.Middleware.ProblemDetails"" Version=""6.5.1"" />";
+        
         return @$"<Project Sdk=""Microsoft.NET.Sdk.Web"">
 
   <PropertyGroup>
@@ -45,7 +49,7 @@ public class WebApiCsProjBuilder
     <PackageReference Include=""Bogus"" Version=""34.0.2"" />
     <PackageReference Include=""EFCore.NamingConventions"" Version=""7.0.2"" />
     <PackageReference Include=""FluentValidation.AspNetCore"" Version=""11.2.2"" />
-    <PackageReference Include=""HeimGuard"" Version=""0.3.0"" />
+    <PackageReference Include=""HeimGuard"" Version=""0.3.0"" />{errorPackages}
     <PackageReference Include=""Mapster"" Version=""7.3.0"" />
     <PackageReference Include=""Mapster.DependencyInjection"" Version=""1.0.0"" />
     <PackageReference Include=""Mapster.EFCore"" Version=""5.1.0"" />